requirements.txt
playwright>=1.46.0,<2
pdfminer.six>=20231228
PyMuPDF>=1.24.10
pdfplumber>=0.11.4
fastapi>=0.112.0
uvicorn[standard]>=0.30.0 